requirements.txt
asgiref==3.7.2
Django==4.2.4
django-debug-toolbar==4.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
Pillow==10.0.0
pytils==0.4.1
sqlparse==0.4.4
